Add files via upload
Assignment two for object oriented programming. Displays a GUI allows a user to connect to an sql database. One way to compile this code is to upload all of these onto blueJ and right click on the planner class and run main.  There is also a make file if you want to compile on the terminal.